[refactor/#14] Converter 디렉토리 생성 및 validation 구현 전 checkpoint
