+++ src/main/java/com/clokey/server/global/config/security/SecurityConfig.java
@@ -32,7 +32,7 @@ public class SecurityConfig {
             "/actuator/prometheus",
             "/ws/**",
             "/topic/**",
-            "**/**"
+            "/**"
     };
 
     @Bean
